Handle double letter edge case.
If you submit a word with double letters, if only one of that letter is
in the solution, then the wordle feedback will indicate that the 2nd
letter is not, and it will get added to letters_not_in_word.  We remove
those so that we do not exclude valid words when we filter the word
list.